refactor login test and resources
This commit removes file loginLocator and moves content
to a new components file to hold the locators and playwright
element handles. Also, renames LoginPage file to LoginController
so improve readability because now this file controls the
login components and perform actions. And, refactor login spec
to match the new files.

Signed-off-by: ltsuda <leonardo.tsuda@icloud.com>